bot/requirements.txt
pydantic==2.7.4
pydantic-settings==2.3.3
python-telegram-bot==21.3
loguru==0.7.2
scikit-learn==1.5.0
numpy==1.26.4
pandas==2.2.2
matplotlib==3.9.0